Adjusted project structure for organization, separated navigation logic from PricingPalApp.kt
