+++ app/src/main/java/com/example/pricingpal/model/dto/ItemDTO.kt
@@ -1,4 +1,4 @@
-package com.example.pricingpal.model
+package com.example.pricingpal.model.dto
 
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.SerialName
